Add test tips on the compass.
Write a new class to represent the compass
in the world space. So the vtkOrientationMarker could
just setMarker to indicate the compass actor.
The GVSCompassActor just act as a marker, and has a
compass with tips N W E S.

GVSClient-22

+++ src/GVSDoc.cpp
@@ -94,7 +94,7 @@ bool GVSDoc::parseProjectByFileName(std::string gvpFullPath)
     }
     gvpFile.close();
 
-    for (int i = 0; i < m_objManager.treeOfGeoObjs.size(); ++i)
+    for (int i = 0; i < (int)(m_objManager.treeOfGeoObjs.size()); ++i)
     {
         QString modelName = m_objManager.treeOfGeoObjs[i].modelName;
         QString gvmFullPath(prjDataDir + modelName + ".gvm");
